Step2 1차 코드 수정
